% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeValues}
\alias{computeValues}
\title{Compute LRRs and BAF values}
\usage{
computeValues(path)
}
\value{
data frame with all computed values
}
\description{
Given the path of a sample  computes  mean and sd for LRRs and  heterozigosity for BAF values per chromosome
}
