% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optimizeClusterFemales}
\alias{optimizeClusterFemales}
\title{Create and optimize Clusters for Females}
\usage{
optimizeClusterFemales(
  dataFrame,
  minPtsRange = c(5, 8, 10, 12, 15, 20, 25, 30, 35, 40, 50, 100),
  interClusProb = 0.5
)
}
\arguments{
\item{dataFrame}{data frame with computed LRR & BAF values}

\item{minPtsRange}{vector with range of values to optimize the minPts parameter}
}
\value{
data frame with all computed values
}
\description{
Given a data frame with mean LRR and BAF values computes SCA clusters for Females
}
