% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotFemaleClusters}
\alias{plotFemaleClusters}
\title{Plot Female Clusters}
\usage{
plotFemaleClusters(dataFrame, optimizeClusterFemalesObj)
}
\arguments{
\item{dataFrame}{data frame with computed LRR & BAF values}

\item{optimizeClusterFemales}{best cluster. Output  from optimizeClusterFemales method}
}
\value{
Plot with male Clusters
}
\description{
Given the optimizeClusterFemales output is plots the results
}
