% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotMaleClusters}
\alias{plotMaleClusters}
\title{Plot Male Clusters}
\usage{
plotMaleClusters(dataFrame, optimizeClusterMalesObj)
}
\arguments{
\item{dataFrame}{data frame with computed LRR & BAF values}

\item{optimizeClusterMales}{best cluster. Output  from optimizeClusterMales method}
}
\value{
Plot with male Clusters
}
\description{
Given the optimizeClusterMales output is plots the results
}
