% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotXandYChr}
\alias{PlotXandYChr}
\title{Plot individual samples X and Y chromosomes}
\usage{
PlotXandYChr(Sample, Name = "Test.png", Type = "png", Save = FALSE)
}
\arguments{
\item{Sample}{data frame with computed LRR & BAF values}

\item{Name}{path/name where to save plot}

\item{Type}{Output file format one from: 'png' (default), 'pdf', 'svg'}
}
\value{
Plot with X and Y chromosomes
}
\description{
Given the optimizeClusterFemales output is plots the results
}
