% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ReadSample}
\alias{ReadSample}
\title{Read sample}
\usage{
ReadSample(
  RawFile = "Test.txt",
  tabix = F,
  skip = 0,
  LCR = NULL,
  PFB = NULL,
  chr = NA,
  SNPList = NULL,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{skip}{number of lines to skip}

\item{chr}{if limit to any specific chromosome}

\item{Rawfile}{path to file}
}
\value{
A matrix of the infile
}
\description{
This function Read a intensity file also tabix
}
