% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulFemales}
\alias{simulFemales}
\title{Simulate female samples}
\usage{
simulFemales(n, mXs, mAs, sXs, sAs, pX, pA)
}
\arguments{
\item{n}{Number of samples to simulate}

\item{mXs}{mean LRR for X chromosome}

\item{mAs}{mean LRR for Autosomal chromosome}

\item{sXs}{sd LRR for X chromosome}

\item{sAs}{sd LRR for Autosomal chromosome}

\item{pX}{deviance in X chromosome for 47XXY carriers}

\item{pA}{deviance in Autosomal chromosome for 47XYY carriers}
}
\value{
data frame with samples (rows)  with Karyotipes, ChrX and Chr-Auts mean LRR values (columns)
}
\description{
Simulate female samples with SCA carriers for clustering
}
