% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulMales}
\alias{simulMales}
\title{Simulate male samples}
\usage{
simulMales(n, mXs, mYs, sXs, sYs, pX, pY)
}
\arguments{
\item{n}{Number of samples to simulate}

\item{mXs}{mean LRR for X chromosome}

\item{mYs}{mean LRR for Y chromosome}

\item{sXs}{sd LRR for X chromosome}

\item{sYs}{sd LRR for Y chromosome}

\item{pX}{deviance in X chromosome for 47XXY carriers}

\item{pY}{deviance in Y chromosome for 47XYY carriers}
}
\value{
data frame with samples (rows)  with Karyotipes, ChrX and ChrY mean LRR values (columns)
}
\description{
Simulate male samples with SCA carriers for clustering
}
