% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SlideWindowMean}
\alias{SlideWindowMean}
\title{Calculate SlideWindow mean}
\usage{
SlideWindowMean(X, N)
}
\arguments{
\item{X}{vector}

\item{N}{length}
}
\value{
A matrix of the infile
}
\description{
This function takes computes the mean over a sequence
}
