% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runPar}
\alias{runPar}
\title{Run in Parallele  computeValues}
\usage{
runPar(paths, cores = 10, tabixF = 0)
}
\arguments{
\item{paths}{Vector containig paths for raw files}

\item{cores}{number of cores}
}
\value{
data frame with all computed values
}
\description{
Applies the computeValues function to run in multiple cores
}
